REFACTOR: Extracted MovieSearchModule from monolithic SearchModule
